third_party/blink/web_tests/external/wpt/css/css-backgrounds/background-332-expected.txt

This is a testharness.js-based test.
[FAIL] background_specified_size
  assert_equals: background specified value for background-size expected "160px auto" but got "160px"
Harness: the test ran to completion.
